Fix naming issues
